requirements.txt
cx-Freeze==5.1.1
Django==1.11.14
pytz==2018.5
RPi.GPIO==0.6.3
channels==1.1.8
asgi_redis
mpu6050-raspberrypi
